fix(restapi): reject out-of-range page instead of truncating it
The movies and series list endpoints typed ?page as uint32 in the spec and
narrowed it to uint16 in the handler, so ?page=65537 silently wrapped to page
1 and served the wrong slice with a 200. Typing the parameter uint16 in the
spec (with the 65535 bound declared, matching ImportPage) makes the generated
binder fail the parse and return 400 for anything wider.

Also aligns schedules' last_duration_ms with storage: ent holds it as uint32,
the spec said int32; the response type now carries the row's type unchanged.

+++ internal/server/restapi/gen.go
@@ -2921,7 +2921,7 @@ type Schedule struct {
 	//
 	// Example: 15m
 	Interval       string     `json:"interval"`
-	LastDurationMs int32      `json:"last_duration_ms"`
+	LastDurationMs uint32     `json:"last_duration_ms"`
 	LastError      *string    `json:"last_error,omitempty"`
 	LastFinishedAt *time.Time `json:"last_finished_at,omitempty"`
 	LastStartedAt  *time.Time `json:"last_started_at,omitempty"`
@@ -3456,7 +3456,7 @@ type SeriesLimit = uint16
 type SeriesLookupQuery = string
 
 // SeriesPage defines model for SeriesPage.
-type SeriesPage = uint32
+type SeriesPage = uint16
 
 // SeriesQuery defines model for SeriesQuery.
 type SeriesQuery = string
@@ -3765,7 +3765,7 @@ type ListImportShowsParamsClassification string
 
 // ListMoviesParams defines parameters for ListMovies.
 type ListMoviesParams struct {
-	Page  *uint32 `form:"page,omitempty" json:"page,omitempty"`
+	Page  *uint16 `form:"page,omitempty" json:"page,omitempty"`
 	Limit *uint16 `form:"limit,omitempty" json:"limit,omitempty"`
 }
